requirements.txt
apache-airflow==2.9.1
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0 